Split up useCustomFetch into composable and type
